remove derived values from db, query instead
there are derived values stored in the database from datafile and
project entities: the detection data range, detection count, and
bounding box. these are set following data file uploads. but they are
not updated in response to datafile deletion and don't take into account
detections that are 'deleted' through the data cleansing feature.
remove these fields and query information when required.

this may lead to degraded performance, particularly on the front page
where we have to query for the detection date and bounding box of each
project; however, we can fix this issue later where it's material.