Support rejection of listing as contributor
Update "added as contributor" email to include rejection link.
Link goes to website confirming they with not to be listed.
Upon confirmation, remove as contributor to project and,
if they are not linked to either a user account or any other projects,
delete the `Person` entity to keep the database tidy.