Read R analysis results from remote file
Now that allow Rserve to be on a remote host, we can no longer write
results to local files read by the web application. So we need to have
the remote R code write to a file, return the path as the result of
the function, and then use RConnection.openFile(fileName) to read it.